refactor(rog-platform): integrate dmi, scsi, and slash modules into rog-platform
