+++ rog-platform/Cargo.toml
@@ -8,6 +8,9 @@ repository.workspace = true
 homepage.workspace = true
 edition.workspace = true
 
+[features]
+dbus = []
+
 [dependencies]
 log.workspace = true
 serde.workspace = true
@@ -18,6 +21,8 @@ inotify.workspace = true
 rusb.workspace = true
 thiserror.workspace = true
 nvml-wrapper.workspace = true
+ron.workspace = true
+libc.workspace = true
 
 [dev-dependencies]
 serde_json = { workspace = true, features = ["preserve_order"] }
