+++ rog-platform/src/lib.rs
@@ -5,6 +5,7 @@ pub mod asus_armoury;
 pub mod backlight;
 pub mod cled;
 pub mod cpu;
+pub mod dmi;
 pub mod error;
 pub mod gpu_pci;
 pub mod hid_raw;
@@ -12,7 +13,11 @@ pub mod keyboard_led;
 pub(crate) mod macros;
 pub mod platform;
 pub mod power;
+pub mod scsi;
+pub mod slash;
 pub mod usb_raw;
+
+pub use dmi::DMIID;
 
 use std::path::Path;
